housekeeping
This updates all of the dependencies and the code itself to use all the
latest paradigms and patterns. This was done by re-building the app on
top of a freshly generated repo created with the latest version of the
Phoenix generator.

It also rips out the Distillery and CircleCI stuff that will no longer
be necessary once deployment happens through Gigalixir instead.