run migrations in pipeline
This sets up the deployment pipeline to run Ecto migrations
automatically whenever a new deployment happens. The new
`bin/migrate.sh` script will start a new Kubernetes job with the newly
built container image, then it will wait for it to complete, print the
logs, and delete it. This should make it so that migrations don't have
to be run manually any more.